Deduplicate literal -> constant lowering
